+++ src/sdk/base/index.ts
@@ -1,8 +1,6 @@
 export { EtherspotWalletAPI } from './EtherspotWalletAPI';
 export { PaymasterAPI } from './PaymasterAPI';
-export { ERC4337EthersSigner } from './ERC4337EthersSigner';
-export { ERC4337EthersProvider } from './ERC4337EthersProvider';
-export { ClientConfig } from './ClientConfig';
+export type { ClientConfig } from './ClientConfig';
 export { HttpRpcClient } from './HttpRpcClient';
 export { DeterministicDeployer } from './DeterministicDeployer';
 export { VerifyingPaymasterAPI } from './VerifyingPaymasterAPI';