+++ src/sdk/common/OperationUtils.ts
@@ -1,13 +1,14 @@
-import { ethers } from 'ethers';
-import { UserOperationStruct } from '../contracts/account-abstraction/contracts/core/BaseAccount';
+import { resolveProperties } from './utils';
+import { BaseAccountUserOperationStruct } from '../types/user-operation-types';
+import { toHex } from 'viem';
 
-export function toJSON(op: Partial<UserOperationStruct>): Promise<any> {
-  return ethers.utils.resolveProperties(op).then((userOp) =>
+export function toJSON(op: Partial<BaseAccountUserOperationStruct>): Promise<any> {
+  return resolveProperties(op).then((userOp) =>
     Object.keys(userOp)
       .map((key) => {
         let val = (userOp as any)[key];
         if (typeof val !== 'string' || !val.startsWith('0x')) {
-          val = ethers.utils.hexValue(val);
+          val = toHex(val);
         }
         return [key, val];
       })
@@ -20,6 +21,6 @@ export function toJSON(op: Partial<UserOperationStruct>): Promise<any> {
       ),
   );
 }
-export async function printOp(op: Partial<UserOperationStruct>): Promise<string> {
+export async function printOp(op: Partial<BaseAccountUserOperationStruct>): Promise<string> {
   return toJSON(op).then((userOp) => JSON.stringify(userOp, null, 2));
 }