+++ src/sdk/types/viem-rpc.ts
@@ -1,0 +1,14 @@
+
+export interface ViemRpcRequestError {
+    details: string;
+    metaMessages: string[];
+    shortMessage: string;
+    name: string;
+    version: string;
+    message: string;
+    cause: {
+      message: string;
+      code: number;
+    };
+    code: number;
+  }